[server] Early request termination and tunable queue capacity
This code change is trying to limit the resource usage when the storage
node is performing quickly enough.
The first feature is called early request termination, and it means when the
request has stayed longer than the configured timeout, it will be aborted
instead of executing them since the requests could be already timed-out on
the caller.
So far, this code change offers a way to configure a timeout threshold per store,
and next step will be pass the user-defined timeout from client to the backend.
New config, and the default value is empty (no threshold is defined):
  /**
   * The request early termination threshold map:
   * The key will be store name, and the value will be the actual threshold.
   * This config is temporary, and in the long run, we will ask Venice Client to pass the actual timeout to the backend.
   */
  public static final String SERVER_STORE_TO_EARLY_TERMINATION_THRESHOLD_MS_MAP = "server.store.to.early.termination.threshold.ms.map";

Metric for this feature:
early_terminated_request_count

The second feature added here is to expose a tunable queue capacity in storage node.
Right now, Storage Node is maintaining a blocking queue to cache all the requests
before executing them. In the worst scenario, this queue could contain
a lot of pending requests when the storage node is slow, and those request objects
could NOT be garbage-collected since there are still strongly referenced.
To mitigate this issue, the tunable queue capacity will bound the pending requests
and when the queue is saturated, the new tasks will be executed in the caller threads
(Netty threads), so that Netty threads won't be able to schedule more tasks (accepting new requests).
When tuning this config, there is a trade-off there: latency vs throughput.
We should consider the following factors:
1. The average latency we expected in the worst scenario;
2. The maximum pending requests we would like to keep in the storage node;

If the queue capacity is too small, the throughput will be affected, and the queue
is too big, the GC behavior could be very bad in the worst scenario.
New configs with default value: Integer.MAX_VALUE:

  /**
   * The following config is used to control the maximum database lookup requests queued, when the queue is full,
   * server will propagate the back pressure to the caller.
   */
  public static final String SERVER_DATABASE_LOOKUP_QUEUE_CAPACITY = "server.database.lookup.queue.capacity";

  /**
   * Check @{@link #SERVER_DATABASE_LOOKUP_QUEUE_CAPACITY} for the explanation.
   * The following config is specifically for compute request.
   */
  public static final String SERVER_COMPUTE_QUEUE_CAPACITY = "server.compute.queue.capacity";

  /**
   * Check the available types in {@literal com.linkedin.venice.config.BlockingQueueType}
   */
  public static final String SERVER_BLOCKING_QUEUE_TYPE = "server.blocking.queue.type";

Available queue types:
  FAIR_BLOCKING_QUEUE,
  LINKED_BLOCKING_QUEUE,
  ARRAY_BLOCKING_QUEUE

If we want to try `ARRAY_BLOCKING_QUEUE`, the queue capacity must be specified.

RB=2032330
BUG=VENG-5692
G=voldemort-reviewers
R=anaberez,smaji,fvillega
A=smaji,fvillega