[router] Introduce the feature flag to control idle connection cleanup
Recently, we have been seeing the latency spike caused by the new connections
setup to storage node.
When the new connection setup is required, the incoming requests will be
delayed since the ssl connection setup will take a fair amount of time.
To keep a stable number of available connections to storage node,
here are steps we could do with this change today:
1. Disable the idle connection cleanup in router;
2. Increase the idle connection cleanup to a big value in storage node;

This code change introduces two new configs in  Router:
/**
   * Whether to enable the cleanup of the idle connections to storage node.
   * Recently, we are seeing latency spike because of new connection setup, and we hope the total available connections will be
   * more stable by disabling the idle connection cleanup.
   * The potential long-term solutions could be connection warm-up for HTTP/1.1 or adopting HTTP/2.
   * Default : false
   */
  public static final String ROUTER_IDLE_CONNECTION_TO_SERVER_CLEANUP_ENABLED = "router.idle.connection.to.server.cleanup.enabled";

  /**
   * The idle threshold for cleaning up the connections to storage node.
   * Default: 3 hours
   */
  public static final String ROUTER_IDLE_CONNECTION_TO_SERVER_CLEANUP_THRESHOLD_MINS = "router.idle.connection.to.server.cleanup.threshold.mins";

The longer-term solution will be:
1. Connection warmup for HTTP/1.1;
2. HTTP/2 (preferred);

RB=2034598

RB=2034598
BUG=VENG-5696
G=voldemort-reviewers
R=anaberez,vgovinda,smaji,apoursam,fvillega
A=smaji