[controller][common] Only log redundant state change notifications
Currenlty we throw a confusing exception when we get a redundant
notification on a state change.  Zk guarantees at most once delivery
of messages, so another process performed the update.  That said, a NOOP
update on a znode is enough to trigger the listeners.  So it's likely
another process can update a znode, succeed but fail on the response
communication, and then retry the operation.  This change just logs the
scenario.

Also changed naming of numberOfPartition to numberOfPartitions as this
was causing me some confusion (could be interpreted as a weird way to
say partitionId).

RB=2034658

RB=2034674
BUG=VENG-5609
G=voldemort-reviewers
R=nithakka,mnhuang,xyin,goliu,zfan,jlliu,xnma,shuhliu,yurzhou,smaji,fvillega,anaberez,sdwu
A=anaberez