[controller] Changes needed to add A/A and push source config to version config
This is the change needed to add the new configuration parameters and
also expose them to being edited by admin tool/controller api.  A call
which tries to update the store to have this config when the store is
not leader follower will automatically fail.

Also applied a fix that was causing some test flakiness in
TestVeniceParentHelixAdmin.

RB=2067357
BUG=VENG-5700
G=voldemort-reviewers
R=zfan,jlliu,shuhliu,xnma,smaji,mnhuang,sdwu,yurzhou,nithakka,xyin,goliu,fvillega,anaberez
A=mnhuang