[controller][common] Do not wait on empty push status when creating a store
It seems polling the children colos doesn't make a ton of sense and is
wreaking some havoc.  A summary of the conversation:

Reverting a store creation that had actually succeeded is an expensive
operation that can lead to some mixed up state.  We also should not poll
on child colo status.  We should instead have the parent trust that it's
child colos will get the admin messages it sends out (though they may be
delayed).  If all messages are delivered the state will eventually
reconcile.  If messages are skipped it will not, and SRE's will need to
be engaged (or some other auto remediation will need to kick in).  The
theory is that this is latter case happens quite rarely.

RB=2049378

RB=2049378
BUG=daas-41325
G=voldemort-reviewers
R=xnma,smaji,mnhuang,yurzhou,goliu,sdwu,zfan,shuhliu,xyin,fvillega,anaberez,jlliu,nithakka
A=sdwu