[router][server] Fixed empty server metrics when conn warming is enabled
This code change has two fixes:
1. Allow negative retry threshold to disable retry;
2. Fixed empty server metrics when conn warming is enabled;

For the 2nd fix, the cause is that when there is a health check request
in the current channel, the `newRequest` flag will never be reset, which means
`isHealthCheck` will never be reset, so no matther the following requests
are health check request or not, it will always be treated as the
health check request and no metrics will be emitted for health check requests.
This fix is to reset the `newRequest` flag after serving any request.

RB=2053408

RB=2053408
BUG=VENG-5766
G=voldemort-reviewers
R=smaji,mnhuang
A=smaji