[samza][integration-test] Fix the test regression in VeniceSystemFactoryTest
Fix the test regression in VeniceSystemFactoryTest#testGetProducer. The
test used to run in a hacky way where we wrote parital update messages
to an store that didn't enable write-compute feature. Later, we had
checks in our schema repository that only monitored derived schema
changes if the stores were write-compute enabled. This led to the test
failure. This RB simply removed write-compute section from the test and
we'll move that part to a new place in next RB.

Additionally, we also add a safety check in the Samza plugin to check
if write-compute is enabled before sending partial update messages.

RB=2069149
R=fvillega,anaberez,shuhliu,mnhuang,xyin,zfan,zpolicze,jlliu,xnma,goliu,yurzhou,nithakka,smaji
A=yurzhou