[server] GC optimization for compute use case and fixed compute latency jump issue when ZK reconnect happens
Previously, for each record, Venice Server would reuse the existing BinaryDecoder
and pass a new InputStream for de-serialization, but the default Avro implementation
of DecoderFactory would allocate a big buffer for each input stream: 8KB.
Say a compute request to storage node contains 100 keys/records, the buffers allocated
by DecoderFactory itself will be 800KB, which is a big overhead.
This code change creates a DecoderFactory with much smaller buffer size: 512 Bytes.
And also, since the PYMK use case is not using chunking/compression, so the input stream
is not needed at all, but just for the code consistency.
This code change introduces a new class: VeniceByteArrayInputStream, so that we could
differentiate the PYMK use case from others, and we don't need to use InputStream at all
since all the bytes are already inside a byte array, and even the regular Avro implementation
won't need to allocate a separate buffer for byte array when reusing the same BinaryDecoder.

In venice-p, this code change has showed a great reduction of GCCount (in some node, more than 50%),
which in return reduces the unnecessary cpu usage and improves the p99 latency greatly (more than 50%)
in stable state.

We noticed that there is a big latency/cpu usage jump when ZK disconnect/reconnect happens, and here is
the detailed reasons (even so, I still don't know why it could cause a persistent much higher GC count/
CPU usage), but this code change does introduce a way to avoid this behavior.
Here is the copy of the description of the new class: StoreValueSchemasCacheService, which will be used
in compute read path:

This class implements the fast value schema/latest value schema lookup with acceptable delay.
The reason to introduce this class is that we found two issues to use {@link HelixReadOnlySchemaRepository} directly
in read compute path:
1. When ZK disconnect/re-connect happens, {@link HelixReadOnlySchemaRepository} will refresh
its local cache, which would cause an increased GC count in read compute since {@link HelixReadOnlySchemaRepository#refresh}
is holding a giant write lock and all the value schema/latest value schema lookups in the read compute requests will be blocked.
The GC count increase is significant (more than doubled in test cluster), which has been causing much higher CPU usage and higher latency;
2. The schema objects returned by {@link HelixReadOnlySchemaRepository} for the same schema
are not always the same object since {@link HelixReadOnlySchemaRepository#refresh()} would always re-create new Schema objects,
which will cause the inefficient de-serializer lookup in {@link SerializerDeserializerFactory}, which will compare the schema objects
to find out the corresponding serializer/de-serializer (for read compute case, de-serializer is the concern). If the schema objects
are not the same, {@link Schema#hashCode()} and {@link Schema#equals(Object)} will be used, in Avro-1.7 or above, {@link Schema#hashCode()} is
optimized to only calculate once if it is read-only, but {@link Schema#equals(Object)} couldn't be avoided.

Here how it works in this class:
1. It maintains a mapping between stores and their value schemas and latest value schema;
2. It will try to reuse the same {@link Schema} object for the same Schema Id within a store since value schema is immutable;
3. It maintains a refresh thread to update the local cache periodically;

In theory, all the schema lookups shouldn't be blocked by invoking the underlying {@link HelixReadOnlySchemaRepository} since in reality,
it will take a fair long time to register a new value schema/latest value schema and start using it in prod, so the periodical schema
refresh should be able to take care of the new value schema/latest value schema discovery.
Since the refresh is async, there is a delay there (at most 1 min), and it should be acceptable because of the previous assumption.

So far, this class only supports value schema by id lookup and latest value schema lookup.

With this change, I have manually introduced a thread to periodically trigger {@link HelixReadOnlySchemaRepository#refresh()},
which is a part of ZK reconnect, and the latency jump never happens.
Also always returning the same Schema object for the same Schema itself will yield another ~10% of improvement on top of this optimization.

With those changes, read compute latency is very stable and it is much lower than the previous release in test cluster. (2.3ms vs 7~8ms).

RB=2061455
BUG=VENG-5854
G=voldemort-reviewers
R=apoursam,smaji,fvillega,zpolicze
A=zpolicze,smaji