[server] fix hybrid store disk quota in SN doesn't reflect the store config
Currently for the hybrid quota, it is only checked with store config in the StoreIngestionTask constructor. Thus if in the beginning the config is set to true and begins ingestion and in the ingestion, then we disable the hybrid quota config, but the ingestion will still use the “true” value.

This rb checks the store config every time before doing enforcement.

RB=2070274
BUG=VENG-5879
G=voldemort-reviewers
R=mnhuang,xyin,goliu
A=fvillega