[etl] fix pattern mismatch issue when enabled more than one store for on-demand ETL
On 04/22 when doing testing with Waterloo on WAR, we found that when enabled more than one store for on-demand ETL, the destination path is mismatched. For example, when only enable one store "RawProfileTermVectors" to on-demand, the deltas are published to https://lva1-waraz01.grid.linkedin.com:8443/hdfs/jobs/veniceetl001/dbchanges/RawProfileTermVectors, which is the correct destination.

But when we enabled two or three store for on-demand, "RawProfileTermVectors" is published to https://lva1-waraz01.grid.linkedin.com:8443/hdfs/jobs/veniceetl001/dbchangesveniceetl001/dbchanges/RawProfileTermVectors. It duplicated the "/dbchangesveniceetl001".

The issue is caused by the class field destinationDirPrefix, when having more than one on-demand stores, destinationDirPrefix keeps adding up the veniceeetl001/dbchanges parts. This rb fixes this issue. This bug was not caught in faro because we didn't enable more than one store for on-demand in faro. Good lessons learnt XD

Also give a new line between current version and future version for metadata file for better readibility: https://lva1-waraz01.grid.linkedin.com:8443/hdfs/jobs/veniceetl001/dbchanges/MemberStandardizedGeo/metadata.txt

RB=2070524
BUG=VENG-5888
G=voldemort-reviewers
R=goliu,shuhliu,smaji,anaberez,fvillega,sdwu,jlliu,xyin,zfan,xnma,nithakka,zpolicze,mnhuang
A=mnhuang