[server] Add versioned stats for monitoring hybrid stores
1. Replaced largest_version_kafka_real_time_buffer_offset_lag and largest_version_number_of_partitions_not_receive_SOBR
with corresponding versioned storage ingestion stats.

2. Added a new rt_topic_offset_lag_over_threshold metric to capture the lag above the user set threshold for setting up alert.

3. Added a new gauge metric ingestion_task_errored_gauge to monitor scenarios when the store ingestion task encounters unexpected
errors and is forced to stop/close. This is important for the current version of hybrid stores because the replica will not go to
error state (thus not triggering any rebalance or error reporting) and the replica will stay stale unless we bounce the server.

The gauge may not reset properly if all partitions are rebalanced to a different SN after the ingestion task errored but before it's
restarted. Very unlikely unless the store version has very few partitions and a rebalance is triggered. The current version's gauge
value is useful even for non-hybrid stores because if the ingestion task died unexpectedly then it will not handle various consumer
actions properly (e.g. UNSUBSCRIBE and RESET_OFFSET).

RB=2079209
BUG=VENG-5819
G=voldemort-reviewers
R=sdwu,mnhuang,goliu,fvillega
A=mnhuang