[server] Fixed mishandled DIV error that causes inconsistency between ingestion task and replica state
reportError's PreNotificationCheck only cares about if EOP is received and not check if the version is the current
version. Current version replicas should not go to ERROR state due to any DIV exceptions since the batch part have
passed DIV before already and RT DIV exceptions are acceptable (due to compaction).

The old handling logic had a hole which causes inconsistency states when the version is the current version but end of
push is not yet received. This caused the replica to go to ERROR state without unsubscribing the faultyPartition from the
ingestion task. As a result we will get the misleading log about consumption will continue and a bunch of skipping message
log since the partition is already errored.

RB=2080017