[common][router] Router implementation of Read path of ZSTD_WITH_DICT.
This RB implements the router implementation of fetching a dictionary from the Storage Node when dictionary compression is ZSTD_WITH_DICT. If the router receives a request for a key but the dictionary for the current version has not been downloaded, it falls back to the previous version. When the router doesn't find a suitable version to fallback to, it will fail the request as a 500 server error.

RB=2085250

RB=2085406
G=voldemort-reviewers
R=zpolicze,xnma,shuhliu,xyin,zfan,sdwu,fvillega,goliu,jlliu,mnhuang,smaji,yurzhou,anaberez
A=goliu