[router][server] connectioin warming for httpasyncclient
This code change introduces a `ClientConnectionWarmingService`, which will in charge of
connection warming for httpasyncclient when running in per-node-client mode.
We are not planning to support connection warming for httpasyncclient running in pooling mode,
since we are planning to migrate to netty client.
This code change doesn't support connection warming for netty client, which we will explore it later.
{@link ClientConnectionWarmingService} will take care of 3 different cases:
1. Router restart, and it will warm up the clients to all the live instances in {@link #startInner()};
2. Storage Node restart, and it will monitor the live instance change and warm up a new client for the new instances;
3. Connection pool maintenance. A {@link #clientConnHealthinessScannerThread} will periodically scan all the
   existing clients and if the number of available connections of some client is below {@link #connectionWarmingLowWaterMark},
   it will try to create a new client and warm it up to {@link #maxConnPerRoutePerClient} connections and replace it;

More details:
For #1, it is quite clear, and if the connection warming of the clients to all the live instances couldn't be accomplished
within {@link #CONNECTION_WARMING_TOTAL_TIMEOUT_MS}, it will throw an exception to let Router fail fast.

For #2, Router will delay the new instance until the timeout threshold: {@link #newInstanceDelayJoinMs} or the connection
warming is done before the timeout. Since connection warming is a best-effort, we would like to prioritize the availability.

For #3, Whenever the {@link #clientConnHealthinessScannerThread} finds a candidate, it will try to create a new client
and warm it up and swap with the old client, and the old client will be gracefully shutdown by the pre-defined delay:
{@link #CLIENT_GRACEFUL_SHUTDOWN_DELAY_IN_MS}. So far, we haven't found a way to do connection warming with the existing
{@link CloseableHttpAsyncClient} reliably without affecting the live traffic.

On storage node, when server gets restarted, a delay was added in this code change to allow Router connection warming finish
to avoid availability issue mentioned in #2.

New configs:
  /**
   * When the available connections in an httpasyncclient is below the low water mark, the connection warming service will try to
   * spin up a new client to replace it.
   * In theory, this config must be lower than  {@link #ROUTER_MAX_OUTGOING_CONNECTION_PER_ROUTE}.
   * default: 60
   */
  public static final String ROUTER_HTTPASYNCCLIENT_CONNECTION_WARMING_LOW_WATER_MARK = "router.httpasyncclient.connection.warming.low.water.mark";

  /**
   * Connection warming executor thread num, default: 6.
   */
  public static final String ROUTER_HTTPASYNCCLIENT_CONNECTION_WARMING_EXECUTOR_THREAD_NUM = "router.httpasyncclient.connection.warming.executor.thread.num";

  /**
   * For the new instance (Storage Node) detected by Router, the following config defines how much delay because of connection warming it could tolerate.
   * If the connection warming takes longer than it, Router will put it in to serve online traffic by creating a new client without connection warming.
   * default: 2 mins
   */
  public static final String ROUTER_HTTPASYNCCLIENT_CONNECTION_WARMING_NEW_INSTANCE_DELAY_JOIN_MS  = "router.httpasyncclient.connection.warming.new.instance.delay.join.ms";

  /**
   * This config is used to control how much time Server will wait for connection warming from Routers.
   * This is trying to avoid availability issue when router connection warming happens when Server restarts.
   * In theory, this config should be equal to or bigger than {@link #ROUTER_HTTPASYNCCLIENT_CONNECTION_WARMING_NEW_INSTANCE_DELAY_JOIN_MS}.
   * Default: 0
   */
  public static final String SERVER_ROUTER_CONNECTION_WARMING_DELAY_MS = "server.router.connection.warming.delay.ms";

New metric:
unhealthy_host_delay_join.OccurrenceRate : track the unhealthy host requests because of the new added delay join behavior

For connection warming behavior, we could look at `connection_pool` metrics to understand the impact.
And if it is necessary to add more metrics to track connection warming behavior, we could add more.

RB=2088173
BUG=VENG-5905
G=voldemort-reviewers
R=mnhuang,smaji
A=mnhuang