[etl] fix data freshness failure for retired store version
The data freshness flow has been failing for a while. The reason is not
that the snapshots are not fresh any more on the day.
Instead, the failure is caused by the flow also checks for retired versions
for a store that whether each day has new snapshots published.e.g.
https://lva1-waraz01.grid.linkedin.com:8443/executor?execid=26911211&job=data-freshness-validator

The rb fixes this by only checking current version snapshots and
potentially future version snapshots.

RB=2086381
BUG=VENG-5884
G=voldemort-reviewers
R=anaberez,jlliu,goliu,nithakka,shuhliu,fvillega,zfan,mnhuang,xyin,zpolicze,xnma,sdwu,smaji
A=mnhuang