[router] PartitionStatusOnlineInstanceFinder should check zk if it has no information for a requested offline push job, and refresh ResourceAssignment
A few things in here.  First, the HelixRoutingDataRepository and the
OnlineInstanceFinder both have their own watchers on zk and can have
their data out of sync with one another depending on the order of
callbacks received.  The onlineInstanceFinder on receipt of a data
change notification, will also refresh the resourceAssignment object to
be certain that everything is up to date before it tries to ascertain
the state of the push job.

The second patch deals with the case where maybe the callback was straight missed.  When a request for store status is made to a venice router, VeniceDelegateMode.java will make a call to our online host finder for a group of hosts.  If it finds that there are no 'hosts' returned from this call, it will mark Scatter with a 'offlineRequest'.  Later, in VeniceDelegateMode on the scatter call, we'll check the number of offlineRequests in the scatter, and if this value is greater then 0, we'll fail the request (we also increment this counter on a failed host healthcheck).  It's unclear to me why we added this check, but the root of the failure seems to be triggered in our PartitionStatusOnlineInstanceFinder class, where the list of push jobs in this class is out of date with zk (it will not be in the hashmap at time of query, and this will bubble up as an error on the request).  So, on map miss, we try to refresh infromation from zk.  This seems to work to fix the test case.

Added a utility function that converts and external view to a partition
assignment object as that seemed useful.  Also rigged it so that we can
update a single routing assignment as opposed to all of them every time
(as that could get costly)

RB=2087164
BUG=VENG-5926
G=voldemort-reviewers
R=fvillega,smaji,yurzhou,mnhuang,nithakka,zfan,anaberez,xnma,xyin,sdwu,goliu,shuhliu,jlliu
A=sdwu,smaji,mnhuang