[controller][server][da-vinci][common][h2v][admin-tool] Integrated scala-free Kafka admin client
Created a new KafkaAdminWrapper interface, with two implementations:
- KafkaAdminClient: the new officially-supported pure-Java Kafka admin client.
- ScalaAdminUtils: the old deprecated Scala admin client.

Renamed the VeniceConsumerFactory interface to KafkaClientFactory, and made it
responsible for instantiating the KafkaAdminWrapper as well. It does so via
reflection so that in the future we could even go so far as to place the Scala
dependency in a separate module.

The backend components (server, controller) default to the old ScalaAdminUtils
for now, while the Da Vinci client defaults to the new KafkaAdminClient. This
can be overridden with a new config:

kafka.admin.class : the fully qualified class name of the KafkaAdminWrapper.

The controller cannot be ported yet to the new Admin because of one last
remaining operation which is not supported in the new Admin. This is the
check against ZK to see if any topic deletions are currently underway. If
desired, this check could be disabled by setting the concurrent topic deletion
policy to false, but that is not configurable at this time.

The server should be compatible with the new Admin (untested).

The TopicManager cut a fair bit of fat, including some of its internal state
and telescopic constructors. Also discovered that the ZkUtils object did not
get closed when the TopicManager got closed, potentially leading to a leak
(though probably only relevant in tests) -- fixed it.

Tweaked the build so that the da-vinci-client module excludes scala, just like
the venice-samza module.

RB=2093265

RB=2093269
BUG=VENG-5940
G=voldemort-reviewers
R=mnhuang,sdwu,goliu,zfan,nithakka,shuhliu,jlliu,xnma,yurzhou,zpolicze,anaberez,xyin,smaji
A=anaberez