[etl] ETL consumer don't have to consume until maxOffset
The "consume" API will be invoked multiple times in extractor, consumer
doesn't have to consume all the way to maxOffset; otherwise, it could
result in out of memory error and eventually fail the task.

This RB fixes the issue by breaking the for loop in "consume" API
once it encounter real data records in the topic.

RB=2106041
BUG=VENG-5992
G=voldemort-reviewers
R=nithakka,goliu,zpolicze,anaberez,zihli,xyin,yurzhou,zhchen,shuhliu,smaji,jlliu,xnma,fvillega,hutran,sdwu,zfan
A=yurzhou