[router] Disabled async start in Router by default
The original async start feature was  introduce by get around of the non-deterministic started order of Venice components
(Controller/Router/Storage Node) in mint test, but it is risky since a successful router restart could not guarantee the
newly started Router is in a healthy state since the async start will terminate the Router process after.

We would like to disable this feature by default and only enable it in the env, which requires this feature, such as dev env.

RB=2118115
BUG=VENG-6029
G=voldemort-reviewers
R=yurzhou,sdwu
A=yurzhou