[common] bump helix-lib version to 1.0.0.50
Bump up to latest version so that if garbage znodes issue happens again, we can dive in
from latest version. Also the version number is aligned with HaaS controller helix-lib version.

In my knowledge, essentially 1.0.0.50 has same API implementations as 1.0.0.2 and was sitting
in Venice PROD for a week.

RB=2132469
G=voldemort-reviewers
R=fvillega,anaberez,nithakka,goliu,jlliu,zfan,sdwu,mnhuang,smaji,zpolicze,xyin,xnma,shuhliu
A=mnhuang