[controller] Catch and handle TopicDoesNotExistException properly during topic truncation
Recently we added a new behavior that blocks on rt topic truncation during store deletion in the admin channel
and actively try to check/update retention to ensure the rt topic is deleted before proceeding. This allowed
for a race condition that the topic happens to be deleted just as we are trying to confirm its retention is
updated to -1. The unhandled TopicDoesNotExistException will trigger unnecessary error reporting and EKG tickets.

RB=2141943
BUG=VENG-6093
G=voldemort-reviewers
R=yurzhou
A=yurzhou,fvillega