[server] LeasderFollower state model notifier
Add an optional latch between Offline to Follower state transition. The
latch is placed if the version state model served is the current
version. (during cluster rebalancing or SN rebouncing) The latch is a
safe guard to prevent Helix "over-rebalancing" the cluster and failing
the read traffic since the rebalancer only looks EV to determine if a
replica is up. (In delay rebalancer, Helix guarantees to only boost 1
replica per partition per time. However, the rebalancer checks the
availability by only looking if replicas' EV aligins with its
IDEALSTATE.)

RB=2143503
BUG=VENG-6076
R=shuhliu,xnma,yurzhou,fvillega,xyin,nithakka,smaji,zfan,zpolicze,jlliu,mnhuang,goliu,anaberez
A=mnhuang