[controller] Fixed leaking ingestion tasks when retiring old ONLINE versions
We used to prevent sending kill message if the version is already ONLINE to guard against inconsistent state during partial
failure scenarios. However, the same guard would also prevent us from sending out the signal to kill the ingestion task when
retiring old store versions. This caused a memory leak of ingestion tasks in the SN for retired ONLINE versions.

Introduced a new parameter to force kill ingestion tasks that are going to be retired or is in an invalid state.

RB=2144604
G=voldemort-reviewers
R=fvillega,zpolicze,goliu,nithakka,yurzhou,jlliu,zfan,anaberez,xnma,sdwu,mnhuang,smaji,shuhliu
A=smaji,goliu