[server] fix hybrid quota for L/F consuming topic
When L/F is enabled, within same StoreIngestionTask a partition can be
either Leader or Standby or InTransition. It's possible that the partition
is consumed from the real time topic instead of version topic depends on
state transitions and consumption speed. When a partition is consuming from RT
topic, the topic name in StoreIngestionTask is still the version topic name.
It doesn't make sense that we pause the consumption for VT when quota is exceeded,
given the real consuming topic is RT.

This rb checks whether leader state is set and the leader topic to be uesd for pausing
topic consuming.
Also renamed the kafkaTopic in StoreIngestionTask to be kafkaVersionTopic. It can't be RT topic
during initializing.

RB=2157869
BUG=VENG-6050
G=voldemort-reviewers
R=nithakka,xnma,smaji,sdwu,xyin,fvillega,jlliu,zpolicze,shuhliu,anaberez,mnhuang,zfan,goliu
A=sdwu