[etl] Do not publish an empty ETL snapshot to users
If a new version become online in between 2 consecutive lumos workflows,
an empty snapshot becomes the latest snapshot for the new version;
in this case, an empty snapshot will be published to users, which
would cause problems if users also don't check the size of the
snapshot.

This RB fixes the issue by skipping empty snapshots in the publisher
workflow.

RB=2160230
BUG=VENG-6159
G=voldemort-reviewers
R=zpolicze,fvillega,yurzhou,xnma,goliu,sdwu,shuhliu,xyin,anaberez,zfan,nithakka,jlliu,smaji
A=yurzhou