server] Clean up customized state in state transition when partition is dropped
When a Venice resource is dropped or a particular partition is moved to a different instance during rebalance, the storage node that previously held this resource(partition) should remove its customized state from Zookeeper, otherwise, there will be stale states. Helix provides an API to delete customized state for a certain partition. This commit implements the cleanup logic for customized state in offline -> dropped state transition to ensure the deletion of stale states. Corresponding tests are modified to reflect this change.

RB=2164478
G=voldemort-reviewers
R=mgao,goliu,sdwu,yurzhou
A=sdwu,yurzhou