[common][test] adds end2end integration test for customized view
This rb adds an end2end integration test for customized view updating and listening.
During putting up the test, debugged a code bug and a test bug:

1. In `HelixCustomizedViewRepository.java`, we need to change to pass in the execution status name,
which is a string. Otherwise, the `equals()` always returns false because `STRING` is not
possible to be equal to `ExecutionStatus`.
2. Same reason above, we need to compare two strings in `TestHelixCustomizedViewRepository`. This was also
the reason that why `TestHelixCustomizedViewRepository` unit test does not catch the above code bug.
3. Added a end2end test `TestHelixCustomizedView` which tests from SN writes customzied states ->
   Helix admin does the aggregation -> Router reads the aggregated customized view

RB=2177938
BUG=VENG-6006
G=voldemort-reviewers
R=xyin,mnhuang,sdwu,zfan,jlliu,xnma,anaberez,goliu,nithakka,mnzhang,shuhliu,fvillega,zpolicze,smaji
A=sdwu