[common]Fix NPE in Push monitor
This can happen if the path read contians no data to start with.  Compare and set will end up throwing an NPE as it will pass a null into the updater.  This patch protects against the null case, and adds in a default object to be overriden should it be null.

RB=2192922
BUG=VENG-6207
G=voldemort-reviewers
R=mnhuang,nithakka,goliu,shuhliu,fvillega,zfan,xyin,jlliu,smaji,anaberez,sdwu,yurzhou,xnma
A=nithakka