[server] avoid state transition happens before partition status notifier is added
This is for bug fix that routers cannot find online storage for several resources and
causes user queries failures.

The reason that we have such failure is that when SN reports the resources states to
notifiers in dispatcher in ingestion, PartitionPushStatusNotifier is not added to dispatcher yet.
So that PartitionPushStatusNotifier missed the resouces states updates which mis-directs routers
to find the wrong hosts to find online storage.

PartitionPushStatusNotifier creation relies on helix manager to get connected, but once
helix manager is connected, we will start state transitions. And from logs in time sequence what happened was
helix manager connected -> state transition kicks off -> then PartitionPushStatusNotifier is added
to notifier dispatcher. This was causing PartitionPushStatusNotifier doesn't get updates for some resources
state transitions.

This rb fixes it by holding on participant model starting creating consumer tasks till notifier is added to dispatcher, so that state transition is not gonna start.

RB=2189798
BUG=VENG-6259
G=voldemort-reviewers
R=mnhuang,goliu,mnzhang,sdwu
A=goliu,sdwu,anaberez