[venice-thin-client] D2ServiceDiscovery.discoverD2Service is vulnerable to ExecutionException
Based on the incident and CompletableFuture's JavaDoc, get() is expected to throw ExecutionException if
the future completed exceptionally. The existing retry logic in D2ServiceDiscovery.discoverD2Service does
not take this into consideration and will fail without retry if an exception is thrown (just like what
happened in the reported incident).

RB=2199011
BUG=VENG-6250
G=voldemort-reviewers
R=yurzhou,zfan,goliu,mnhuang,shuhliu,xnma,jlliu,zpolicze,sdwu,nithakka,smaji,anaberez,fvillega
A=anaberez