[test] Test suite hardening
Audited all direct usage of ServiceFactory#getKafkaBroker() and fixed
the ones that were not guaranteed to close.

Also added auditing code to the ProcessWrapper class so that it can
detect (at GC time and at JVM shutdown time) if it did not get closed,
in which case it logs the stack indicating where it was constructed,
and then attempts to close itself.

RB=2203672
BUG=VENG-6320
G=voldemort-reviewers
R=smaji,xyin,yurzhou,mnhuang,zfan,shuhliu,goliu,zpolicze,anaberez,xnma,nithakka,jlliu,sdwu
A=yurzhou