[controller][admin-tool][compat]Store backup version retention policy
This code change adds a retention policy for backup version in store level:
backupVersionRetentionMs
If it is not specified, it will use the configured retention policy in system level.
How does it work?
1. The store metadata will have a new field: `latestVersionPromoteToCurrentTimestamp` to
   record the timestamp when a new version is promoted to the current version.
2. If the difference between current timestamp and the `latestVersionPromoteToCurrentTimestamp`
   is bigger than the configured backup version retention policy, the backup versions will be
   eligible for cleanup.
3. `StoreBackupVersionCleanupService` in Child Controller will be in charge of cleaning up the
   staled backup vesions, and it will apply a hard-coded delay if the retention policy is too short
   to accomodate the potential delay from Controller to Routers for current version propagation.

This code change adds the following config in Child Controller:
  /**
   * The following config is to control the default retention time in milliseconds if it is not specified in store level.
   * default: 1 week
   */
  public static final String CONTROLLER_BACKUP_VERSION_DEFAULT_RETENTION_MS = "controller.backup.version.default.retention.ms";

  /*
   * The following config is to control whether to enable backup version cleanup based on retention policy or not.
   * default: false
   */
  public static final String CONTROLLER_BACKUP_VERSION_RETENTION_BASED_CLEANUP_ENABLED = "controller.backup.version.retention.based.cleanup.enabled";

Operationally-wise, before enabling this feature in prod, we need to send an announcement to all Venice users.
If there is no explicit requirement, the backup version will only be kept for the configured time period (1 week if we all agree with this number).
If some customers would like to have different retention policies (shorter or longer), we could tune it via admin-tool.
Ideally, in the future, we could expose it in Nuage, so that each customer could decide it when creating a new Nuage store.

RB=2211194
BUG=VENG-6165
R=fvillega,mnhuang,smaji,xnma,sdwu,jlliu,nithakka,zpolicze,anaberez,xyin,yurzhou,zfan,shuhliu
A=smaji