[controller] Add compaction to real time topics by default if write compute is not enabled.
This sets compaction to true for newly created real time topics for stores, as well as turns compaction off on an RT topic for a store should write compute get enabled, or if write compute is enabled at time of real time topic creation.

This also adds a new controller api that enables ONLY ADMINS to be able to set the compaction policy for topics.  The intention of this is to use it to backfill those stores with real time topics with a compaction policy.  Right now the api does not check if write compute is enabled for a store or not.  We could add that protection if we desired it.  My initial preference was to not preclude that possibility down the road.  But it might be better safe then sorry.

RB=2208225
G=voldemort-reviewers
R=nithakka,smaji,shuhliu,xnma,fvillega,zfan,goliu,jlliu,mnhuang,xyin,anaberez,sdwu,yurzhou
A=mnhuang